requirements.txt
pytesseract
pdf2image
Pillow
img2table
opencv-contrib-python-headless
transformers
sentencepiece
protobuf
torch
pyyaml
tensorboard
